Queue indicator changed ( save mean queue for simulation) .Changed wilconxon test to be two way : negative (lose), positive (win), and if neutral(win).Heatmap queue is same parameters that the RT heatmap.
