+++ fire_emergency_simulation/fire_emergency_OG/globals.py
@@ -1,0 +1,11 @@
+class _Globals:
+    def __init__(self):
+        self.set_index = 0
+        self.replication_index = 0
+        self.interval_index = 0
+        self.total_services_index = 0
+        self.keep_same_excel_file = 0
+        self.folder_path = ""
+
+
+globs = _Globals()
